[Feat/#376] 상품 상세 리포트 API

See merge request s11-final/S11P31S103!86

+++ backend/gs-anomaly-detecion-backend/src/main/java/com/ssafy/s103/domain/anomalyproduct/application/repository/AnomalyProductRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-public interface AnomalyProductRepository extends JpaRepository<AnomalyProduct, Integer>, AnomalyProductCustomRepository {
+public interface AnomalyProductRepository extends JpaRepository<AnomalyProduct, Long>, AnomalyProductCustomRepository {
     int countByStatus(AnomalyStatus status);
     
     @Query("SELECT new com.ssafy.s103.domain.batch.dto.response.BatchResponse(" +
